# wal: use page buffered writer for writing records