Fixes #6271
 
/cc @gyuho