@gyuho 2016-08-30
Ok I will try this branch in failpoints cluster.