@xiang90 2016-08-30
LGTM. Have you tried to run the benchmark in wal pkg to see if there is any performance gain?