@heyitsanthony 2016-08-30
@xiang90 benchmark found a bug in the page writer 👍 

Master:
```
BenchmarkWrite100EntryWithoutBatch-4    	     200	  10144426 ns/op	   0.01 MB/s
BenchmarkWrite100EntryBatch10-4         	    2000	    943462 ns/op	   0.11 MB/s
BenchmarkWrite100EntryBatch100-4        	   10000	    116260 ns/op	   0.93 MB/s
BenchmarkWrite100EntryBatch500-4        	   50000	     28449 ns/op	   3.80 MB/s
BenchmarkWrite100EntryBatch1000-4       	  100000	     15095 ns/op	   7.15 MB/s
BenchmarkWrite1000EntryWithoutBatch-4   	     100	  10645949 ns/op	   0.09 MB/s
BenchmarkWrite1000EntryBatch10-4        	    2000	   1016150 ns/op	   0.99 MB/s
BenchmarkWrite1000EntryBatch100-4       	   10000	    144891 ns/op	   6.96 MB/s
BenchmarkWrite1000EntryBatch500-4       	   50000	     42404 ns/op	  23.79 MB/s
BenchmarkWrite1000EntryBatch1000-4      	   50000	     28502 ns/op	  35.40 MB/s
```

This patch:
```
BenchmarkWrite100EntryWithoutBatch-4    	     100	  10629690 ns/op	   0.01 MB/s
BenchmarkWrite100EntryBatch10-4         	    2000	    963754 ns/op	   0.11 MB/s
BenchmarkWrite100EntryBatch100-4        	   10000	    107374 ns/op	   1.01 MB/s
BenchmarkWrite100EntryBatch500-4        	  100000	     23026 ns/op	   4.69 MB/s
BenchmarkWrite100EntryBatch1000-4       	  200000	     11675 ns/op	   9.25 MB/s
BenchmarkWrite1000EntryWithoutBatch-4   	     100	  10485759 ns/op	   0.10 MB/s
BenchmarkWrite1000EntryBatch10-4        	    2000	    997684 ns/op	   1.01 MB/s
BenchmarkWrite1000EntryBatch100-4       	   10000	    121113 ns/op	   8.33 MB/s
BenchmarkWrite1000EntryBatch500-4       	   50000	     25324 ns/op	  39.84 MB/s
BenchmarkWrite1000EntryBatch1000-4      	  100000	     13901 ns/op	  72.58 MB/s
```