@gyuho 2016-08-30
I can confirm that this fixes https://github.com/coreos/etcd/issues/6271 since it used to fail with less than 3 rounds. Now it's over 10 rounds. Thanks!